requirements.txt
pytorch==2.1.0+cu121
sentencepiece
requests
wandb
tensorboard
tqdm